refactor(cli-better-auth): rename setup-key route to api-key, device dirs to cli-auth
- Move POST /api/cli/setup-key -> /api/cli/api-key; the endpoint mints a
  project API key for the session user and isn't intrinsically setup-bound.
- Rename frontend/components/device -> components/cli-auth and
  frontend/lib/actions/device -> lib/actions/cli-auth (CLI auth approval UI +
  actions). The /device route, device_code table, and BetterAuth /api/auth/device/*
  stay (RFC 8628 protocol-bound).
- Update all import sites, the CLI client URL, and CLAUDE.md.
