+++ CLAUDE.md
@@ -298,7 +298,7 @@ Keep comments short. Don't write multi-paragraph rationale blocks — a single t
 
 ## CLI authentication (BetterAuth Device Flow → user-scoped tokens)
 
-- The CLI stores **BetterAuth tokens** (user-scoped), NOT a project API key. `lmnr-cli login` walks BetterAuth's native `deviceAuthorization` flow end-to-end (no custom poll endpoint). `lmnr-cli setup` does the same login AND additionally mints a project API key that it writes ONLY to `./.env` (never into `credentials.json`) via our `POST /api/cli/setup-key` endpoint.
+- The CLI stores **BetterAuth tokens** (user-scoped), NOT a project API key. `lmnr-cli login` walks BetterAuth's native `deviceAuthorization` flow end-to-end (no custom poll endpoint). `lmnr-cli setup` does the same login AND additionally mints a project API key that it writes ONLY to `./.env` (never into `credentials.json`) via our `POST /api/cli/api-key` endpoint.
 - **Token model:**
   - **Access token** = a 15m EdDSA JWT minted by `GET /api/auth/token` (with the session token as bearer) → response `{ token }`. Verifiable via `/api/auth/jwks`. This is what app-server will eventually validate statelessly (JWKS). App-server JWT validation is a deliberate LATER phase — the CLI's JWT is NOT accepted by app-server yet, so `lmnr-cli sql`/ingestion against app-server with the JWT does not work this iteration.
   - **Refresh token** = the BetterAuth **session token**, returned as `access_token` from `POST /api/auth/device/token`. Durable (session `expiresIn` default 7 days, rolling). Refresh op = re-call `GET /api/auth/token` with the session token → fresh JWT, no browser.
@@ -308,8 +308,8 @@ Keep comments short. Don't write multi-paragraph rationale blocks — a single t
   - `POST /api/auth/device/token` — BetterAuth's NATIVE token endpoint. CLI polls here; on approval returns `{ access_token (=session token), expires_in, ... }`. We no longer own a custom poll endpoint.
   - `GET /api/auth/token` (session bearer) → `{ token }` 15m JWT.
   - `GET /api/auth/get-session` (session bearer) → `{ user: { id, email } }` for profile metadata.
-- **`POST /api/cli/setup-key`** (`app/api/cli/setup-key/route.ts`) — `setup`-only. Auth via session bearer (`auth.api.getSession`). Body (zod `.strict()`): `{ projectId?, deviceName? }`. With `projectId` → membership-check + mint. Without → 1 project mints; >1 returns `400 {error:"project_ambiguous", projects:[{id,name,workspaceName}]}`; 0 returns `400 {error:"no_projects"}`. Returns `{apiKey, apiKeyId, projectId, projectName, workspaceId, workspaceName}`. The minted key goes to `./.env` only.
-- **The `bearer()` plugin is required.** `auth.api.getSession({ headers })` only accepts cookie-borne sessions by default; without `bearer()` an `Authorization: Bearer <session-token>` header is ignored. The device-flow `access_token` IS a session token, so the bearer plugin is what lets it round-trip through `getSession()` (used by `/api/auth/token`, `/api/auth/get-session`, and `/api/cli/setup-key`).
+- **`POST /api/cli/api-key`** (`app/api/cli/api-key/route.ts`) — mints a project API key for the session user (the `setup` flow is the only caller today, but the endpoint is generic, not setup-bound). Auth via session bearer (`auth.api.getSession`). Body (zod `.strict()`): `{ projectId?, deviceName? }`. With `projectId` → membership-check + mint. Without → 1 project mints; >1 returns `400 {error:"project_ambiguous", projects:[{id,name,workspaceName}]}`; 0 returns `400 {error:"no_projects"}`. Returns `{apiKey, apiKeyId, projectId, projectName, workspaceId, workspaceName}`. The minted key goes to `./.env` only.
+- **The `bearer()` plugin is required.** `auth.api.getSession({ headers })` only accepts cookie-borne sessions by default; without `bearer()` an `Authorization: Bearer <session-token>` header is ignored. The device-flow `access_token` IS a session token, so the bearer plugin is what lets it round-trip through `getSession()` (used by `/api/auth/token`, `/api/auth/get-session`, and `/api/cli/api-key`).
 - **CLI credentials shape.** `~/.config/lmnr/credentials.json` (mode 0600, **keyed by `userId`**, version 1) stores tokens only: `sessionToken` (refresh), `accessToken` (the 15m JWT), `accessTokenExpiresAt` (from JWT `exp`), `sessionExpiresAt?` (from device token `expires_in`), `issuer`, `baseUrl`, `userEmail?`, `userId`, `createdAt`, `lastUsedAt?`. NO project API key, NO `apiKeyId`. `resolveAuth` refreshes the JWT when within ~30s of expiry (re-mint via `/api/auth/token`); a 401 means the session expired → "run `lmnr-cli login`". The returned JWT bearer is not yet accepted by app-server (later phase).
 - **Logout** best-effort POSTs `/api/auth/sign-out` (session bearer) to revoke the session server-side, then deletes local creds; `--all` loops all profiles.
 - **App-server later phase (two-flow validation):** the eventual design is JWT-via-JWKS for the CLI (user-scoped — so app-server must derive the project from the request path / an explicit `--project` arg, since the JWT carries no project) and project API key for the SDK. Not implemented here.
