+++ frontend/lib/actions/cli-auth/scope.ts
