+++ frontend/tests/test-device-scope.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import { describe, it } from "node:test";
 
-import { buildProjectScope, parseProjectFromScope } from "@/lib/actions/device/scope";
+import { buildProjectScope, parseProjectFromScope } from "@/lib/actions/cli-auth/scope";
 
 const UUID = "123e4567-e89b-42d3-a456-426614174000";
